Use latest rpm instead of the one from Fedora
The Fedora image is only updated once in a while, let's follow the
latest development by using upstream repository.

Note: Currently the F27 repo is not yet available so Fedora one will be
used. But once the F27 repo is deployed this receipt will start using it
instead of the stock one.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>